@@ -5,7 +5,7 @@ describe TweetsController do
     it "returns the default tweet for a bottle" do
       bottle = create(:alvento_sauvignonblanc_bottle_drank)
 
-      get :bottle, :bottle_id => bottle.id, :format => :json
+      get :bottle, params: {:bottle_id => bottle.id }, :format => :json
 
 #      puts response.body
 
@@ -17,7 +17,7 @@ describe TweetsController do
     it "returns empty for a bottle that should not be tweeted" do
       bottle = create(:imocali_vernaccia_bottle_now)
 
-      get :bottle, :bottle_id => bottle.id, :format => :json
+      get :bottle, params: {:bottle_id => bottle.id}, :format => :json
 
       expect(response).to be_success
       expect(response.body).not_to include(bottle.wine.year.to_s)
@@ -31,7 +31,7 @@ describe TweetsController do
 
       wine = create(:kenwood_merlot)
 
-      post 'tweet_bottle', :wine_id => wine.id
+      post 'tweet_bottle', params: {:wine_id => wine.id }
 
       assert_redirected_to wine_path(assigns(:wine))
     end
@@ -39,7 +39,7 @@ describe TweetsController do
     it "returns failure when not authenticated" do
       wine = create(:kenwood_merlot)
 
-      post 'tweet_bottle', :wine_id => wine.id
+      post 'tweet_bottle', params: { :wine_id => wine.id }
 
       expect(response.response_code).to eq(401)
     end
