@@ -8,8 +8,9 @@ WineGuide::Application.configure do
   config.cache_classes = true
 
   # Configure static asset server for tests with Cache-Control for performance
-  config.serve_static_files = true
-  config.static_cache_control = "public, max-age=3600"
+  config.public_file_server.enabled = true
+
+  config.public_file_server.headers = { 'Cache-Control' => 'public, max-age=3600' }
 
   config.eager_load = true;
 
